backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
gtts
pydantic
python-dotenv
PyPDF2
httpx
pdfminer.six
python-docx

